Resolved invalid suggestion issue.

+++ app/src/main/java/com/fafadiatech/newscout/activity/DetailNewsDataActivity.kt
@@ -99,7 +99,7 @@ class DetailNewsDataActivity : BaseActivity(), VerticalViewPager.SwiperListener 
 
             for (i in 0 until list.size) {
                 var sourceItem = list.get(i)
-                var detailListItem = DetailNewsData(sourceItem.id, sourceItem.title, sourceItem.source, sourceItem.category, sourceItem.source_url, sourceItem.cover_image, sourceItem.blurb!!, sourceItem.published_on, 2, 0, sourceItem.article_score.toFloat())
+                var detailListItem = DetailNewsData(sourceItem.id, sourceItem.title, sourceItem.source, sourceItem.category, sourceItem.source_url, sourceItem.cover_image, sourceItem.blurb!!, sourceItem.published_on, 2, 0, sourceItem.article_score)
                 detailList.add(detailListItem)
             }
